Fix: added fake_data_generator, added requirements, read.me, updated .gitignore

+++ kitchen/models.py
@@ -4,7 +4,7 @@ from django.db import models
 
 
 class Cook(AbstractUser):
-    years_of_experience = models.PositiveIntegerField(default=0)
+    years_of_experience = models.PositiveSmallIntegerField(default=0)
 
     def __str__(self):
         return f"{self.username} - {self.years_of_experience} - {self.email}"
